Register views if the `messaging` app
