+++ messaging/views/__init__.py
@@ -1,0 +1,4 @@
+from .contact import *
+from .contact_assignment import *
+from .contact_role import *
+from .email import *
